// backend/src/types/payment.ts
export enum PaymentMethod {
  STRIPE = 'STRIPE',
  PAYPAL = 'PAYPAL',
  APPLE_PAY = 'APPLE_PAY',
  GOOGLE_PAY = 'GOOGLE_PAY',
  AMAZON_PAY = 'AMAZON_PAY',
  BKASH = 'BKASH',
  NAGAD = 'NAGAD',
  ROCKET = 'ROCKET',
  BITCOIN = 'BITCOIN',
  ETHEREUM = 'ETHEREUM',
  MANUAL_CREDIT_CARD = 'MANUAL_CREDIT_CARD',
}